create final render

+++ Animation.java
@@ -11,22 +11,25 @@ import java.text.NumberFormat;
 
 public class Animation {
     private final static NumberFormat PF = NumberFormat.getPercentInstance();
-    private double centerPointX, centerPointY, scale, branchOffsetAngle, startLength, angle, decrementer, startHue,
+    private double centerPointX, centerPointY, scale, startLength, angle, decrementer, startHue,
             startStroke, endstroke;
-    private int numStartBranches, numRepotitions, numframes;
-    private boolean drawOnChange = false, outputToList = true, drawRendering = true;
-    private ArrayList<Lines> lines = new ArrayList<Lines>();
+    private int numRepotitions, numframes, frameRate;
+    private boolean drawOnChange = false, outputToList = true, drawRendering = true, HD = false;
+    private ArrayList<Branch> branches = new ArrayList<Branch>();
     private ArrayList<Image> frames = new ArrayList<Image>();
     private ArrayList<String> imgFileNames = new ArrayList<String>();
     private Canvas2D canvas;
     public Set set = new Set();
     public Get get = new Get();
-    private double[] centerPointXArr, centerPointYArr, scaleArr, branchOffsetAngleArr, startLengthArr, angleArr, decrementerArr, startHueArr,
+    private double[] centerPointXArr, centerPointYArr, scaleArr, startLengthArr, angleArr, decrementerArr, startHueArr,
             startStrokeArr, endstrokeArr;
+    private ArrayList<Branch> startBranches = new ArrayList<Branch>();
 
     public static enum Variable {
-        centerPointX, centerPointY, scale, branchOffsetAngle, startLength, angle, decrementer, startHue, startStroke, endstroke
+        centerPointX, centerPointY, scale, startLength, angle, decrementer, startHue, startStroke, endstroke
     }
+
+    private long generate, draw;
 
     public Animation(Canvas2D canvas, double centerPointX, double centerPointY, double scale, int numStartBranches,
             double branchOffsetAngle, double startLength, double angle, double decrementer, int numRepotitions,
@@ -34,8 +37,6 @@ public class Animation {
         set.centerPointX(centerPointX);
         set.centerPointY(centerPointY);
         set.scale(scale);
-        set.numStartBranches(numStartBranches);
-        set.branchOffsetAngle(branchOffsetAngle);
         set.startLength(startLength);
         set.angle(angle);
         set.decrementer(decrementer);
@@ -46,19 +47,23 @@ public class Animation {
         set.canvas(canvas);
     }
 
-    public void render() {
-        String[] arr = new String[imgFileNames.size()];
-        imgFileNames.toArray(arr);
-        File outputFile = new File("render.gif");
-        try {
-            outputFile.delete();
-            outputFile.createNewFile();
-            Giffer.generateFromFiles(arr, outputFile.getPath(), 2, true);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        for(String fileName : imgFileNames) {
-            (new File(fileName)).delete();
+    public void render(double lengthSec) {
+        if(outputToList) {
+            frameRate = (int) (float) (1000 / (numframes / lengthSec));
+        }else {
+            String[] arr = new String[imgFileNames.size()];
+            imgFileNames.toArray(arr);
+            File outputFile = new File("render.gif");
+            try {
+                outputFile.delete();
+                outputFile.createNewFile();
+                Giffer.generateFromFiles(arr, outputFile.getPath(), (int) (float) (100 / (numframes / lengthSec)), true);
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            for(String fileName : imgFileNames) {
+                (new File(fileName)).delete();
+            }
         }
     }
 
@@ -68,13 +73,17 @@ public class Animation {
         transform.translate((centerPointX - canvas.getResolution().getWidth() / 2) * scale + canvas.getResolution().getWidth() / 2, (centerPointY - canvas.getResolution().getHeight() / 2) * scale + canvas.getResolution().getHeight() / 2);
         transform.scale(scale, scale);
         CTX.transform(transform);
-        generateLines();
-        for (int j = lines.size() - 1; j >= 0; j--) {
-            Lines line = lines.get(j);
+        generatebranches();
+        long start = System.currentTimeMillis();
+        for (int j = branches.size() - 1; j >= 0; j--) {
+            Branch line = branches.get(j);
             CTX.setColor(line.getColor());
+            if(j < startBranches.size()) CTX.setColor(calculateColor(-1));
             CTX.setStroke(new BasicStroke((int) (float) line.getStrokeWidth(), BasicStroke.CAP_ROUND, BasicStroke.JOIN_ROUND));
             CTX.draw(line.getLine());
         }
+        generate += (System.currentTimeMillis() - start);
+        start = System.currentTimeMillis();
         if (outputToList)
             frames.add(canvas.getImage());
         else {
@@ -86,33 +95,22 @@ public class Animation {
                 e.printStackTrace();
             }
         }
+        draw += (System.currentTimeMillis() - start);
         if(drawRendering) canvas.draw();
     }
 
-    public void generateLines() {
+    public void generatebranches() {
         int lastIteration = 0;
         double length = startLength;
-        lines.clear();
-        for (int i = 0; i < numStartBranches; i++) {
-            lines.add(new Lines(new Point2D.Double(0, 0),
-                    new Point2D.Double(
-                            -startLength * Math.cos(branchOffsetAngle + (2 * Math.PI * i) / numStartBranches),
-                            -startLength * Math.sin(branchOffsetAngle + (2 * Math.PI * i) / numStartBranches)),
-                    startStroke, Color.getHSBColor(calculateColor(-1) / 360f % 1, 1.0f, 1.0f), 0));
-        }
+        branches.clear();
+        branches.addAll(startBranches);
         for (int i = 0; i < numRepotitions; i++) {
-            int max = lines.size();
+            int max = branches.size();
             for (int j = lastIteration; j < max; j++) {
-                Lines line = lines.get(j);
+                Branch line = branches.get(j);
                 Point2D initialPos = line.getP2();
-                lines.add(new Lines(initialPos, getPoint(initialPos, line.getAngle() + angle, length),
-                        map(i, 0, numRepotitions, startStroke, endstroke),
-                        Color.getHSBColor(calculateColor(i) / 360f % 1, 1.0f, 1.0f),
-                        i));
-                lines.add(new Lines(initialPos, getPoint(initialPos, line.getAngle() - angle, length),
-                        map(i, 0, numRepotitions, startStroke, endstroke),
-                        Color.getHSBColor(calculateColor(i) / 360f % 1, 1.0f, 1.0f),
-                        i));
+                branches.add(new Branch(initialPos, getPoint(initialPos, line.getAngle() + angle, length), map(i, 0, numRepotitions, startStroke, endstroke), calculateColor(i)));
+                branches.add(new Branch(initialPos, getPoint(initialPos, line.getAngle() - angle, length), map(i, 0, numRepotitions, startStroke, endstroke), calculateColor(i)));
             }
             length *= decrementer;
             lastIteration = max;
@@ -120,10 +118,11 @@ public class Animation {
     }
 
     public boolean initAnimation(int frameCount) {
+        if(!HD) frameCount /= 2;
+        if(frameCount == 0) frameCount = 1;
         if(frameCount <= 0) return false;
         numframes = frameCount;
         angleArr = new double[numframes];
-        branchOffsetAngleArr = new double[numframes];
         centerPointXArr = new double[numframes];
         centerPointYArr = new double[numframes];
         decrementerArr = new double[numframes];
@@ -134,7 +133,6 @@ public class Animation {
         startStrokeArr = new double[numframes];
         for(int i = 0 ; i < numframes; i++) {
             angleArr[i] = angle;
-            branchOffsetAngleArr[i] = get.branchOffsetAngle();
             centerPointXArr[i] = get.centerPointX();
             centerPointYArr[i] = get.centerPointY();
             decrementerArr[i] = get.decrementer();
@@ -147,19 +145,20 @@ public class Animation {
         return true;
     }
 
-    public boolean addAnimation(Variable attribute, int startFrame, int endFrame, double startValue, double endValue) {
+    public void addAnimation(Variable attribute, int startFrame, int endFrame, double startValue, double endValue) {
+        if(!HD) {
+            startFrame /= 2;
+            endFrame /= 2;
+        }
+        endFrame = Math.min(endFrame, numframes);
+        startFrame = Math.max(startFrame, 0);
         endFrame--;
-        if(endFrame <= startFrame) return false;
-        if(endFrame < 0) return false;
-        if(endFrame >= numframes) return false;
+        endFrame = Math.max(endFrame, 0);
         for(int i = startFrame; i <= endFrame; i++) {
             double value = map(i, startFrame, endFrame, startValue, endValue);
             switch(attribute) {
                 case angle:
                     angleArr[i] = value;
-                    break;
-                case branchOffsetAngle:
-                    branchOffsetAngleArr[i] = value;
                     break;
                 case centerPointX:
                     centerPointXArr[i] = value;
@@ -185,22 +184,24 @@ public class Animation {
                 case startStroke:
                     startStrokeArr[i] = value;
                     break;
-                default:
-                    return false;
             }
         }
-        return true;
     }
 
     public void animate() {
+        boolean hasDrawOnChange = drawOnChange;
+        drawOnChange = false;
+        long start = System.currentTimeMillis();
         for(int i = 0; i < numframes; i++) {
             String str = "|";
             for (int j = 0; j < 25; j++)
-                str += ((float) i / numframes * 25 < j ? " " : "#");
-            str += "| (" + (i + 1) + " / " + numframes + ") (" + PF.format((double) (i + 1) / numframes) + ")\r";
+                str += (Math.floor((float) i / numframes * 25) < j ? " " : "=");
+            long elapsed = System.currentTimeMillis() - start;
+            long millisPerFrame = (elapsed / (i + 1));
+            long Queued = (numframes - 1 - i);
+            str += "| (" + (i + 1) + " / " + numframes + ") (" + PF.format((double) (i + 1) / numframes) + ") (" + getTime(Queued * millisPerFrame) + ")             \r";
             System.out.print(str);
             set.angle(angleArr[i]);
-            set.branchOffsetAngle(branchOffsetAngleArr[i]);
             set.centerPointX(centerPointXArr[i]);
             set.centerPointY(centerPointYArr[i]);
             set.decrementer(decrementerArr[i]);
@@ -211,7 +212,26 @@ public class Animation {
             set.startStroke(startStrokeArr[i]);
             generateFrame();
         }
-        System.out.println();
+        long totalTime = System.currentTimeMillis() - start;
+        System.out.println("\nTotal Time: " + getTime(totalTime) + "\nGenerate Time: " + getTime(generate) + "\t" + PF.format(((double) generate) / totalTime) + "\nDraw Time: " + getTime(draw) + "\t\t" + PF.format(((double) draw) / totalTime));
+        drawOnChange = hasDrawOnChange;
+    }
+
+    private static String getTime(long time) {
+        long millis = time % 1000;
+        long second = (time / 1000) % 60;
+        long minute = (time / (1000 * 60)) % 60;
+        long hour = (time / (1000 * 60 * 60)) % 24;
+
+        return String.format("%02d:%02d:%02d.%d", hour, minute, second, millis);
+    }
+
+    public void addBranch(Point2D point, double angle) {
+        startBranches.add(new Branch(point, Math.toRadians(angle), -startLength, startStroke, calculateColor(-1)));
+    }
+
+    public void addBranch(double x, double y, double angle) {
+        addBranch(new Point2D.Double(x, y), angle);
     }
     
     private Point2D getPoint(Point2D point, double angle, double distance) {
@@ -219,11 +239,15 @@ public class Animation {
     }
 
     private double map(double x, double x1, double x2, double y1, double y2) {
-        return (((y2 - y1) / (x2 - x1)) * (x - x1) + y1);
+        return (isEqualTo(x1, x2) ? y1 : (((y2 - y1) / (x2 - x1)) * (x - x1) + y1));
     }
 
-    private float calculateColor(int index) {
-        return (float) map((index + 2), 1, numRepotitions, startHue, startHue + 360);
+    private static boolean isEqualTo(double num1, double num2) {
+        return (Math.abs(num1 - num2) < 0.000001);
+    }
+
+    private Color calculateColor(int index) {
+        return Color.getHSBColor((float) map((index + 2), 1, numRepotitions, startHue, startHue + 360) / 360f % 1, 1.0f, 1.0f);
     }
 
     public class Set {
@@ -241,12 +265,6 @@ public class Animation {
 
         public void scale(double scale) {
             Animation.this.scale = scale;
-            if (drawOnChange)
-                generateFrame();
-        }
-
-        public void branchOffsetAngle(double branchOffsetAngle) {
-            Animation.this.branchOffsetAngle = Math.toRadians(branchOffsetAngle);
             if (drawOnChange)
                 generateFrame();
         }
@@ -287,10 +305,11 @@ public class Animation {
                 generateFrame();
         }
 
-        public void numStartBranches(int numStartBranches) {
-            Animation.this.numStartBranches = numStartBranches;
-            if (drawOnChange)
-                generateFrame();
+        public void startBranches(int numStartBranches, int branchOffsetAngle) {
+            double angle = Math.toRadians(branchOffsetAngle);
+            for (int i = 0; i < numStartBranches; i++) {
+                startBranches.add(new Branch(new Point2D.Double(0, 0), (angle + (2 * Math.PI * i / numStartBranches)), -startLength, startStroke, calculateColor(-1)));
+            }
         }
 
         public void numRepotitions(int numRepotitions) {
@@ -322,6 +341,10 @@ public class Animation {
         public void drawRendering(boolean drawRendering) {
             Animation.this.drawRendering = drawRendering;
         }
+
+        public void HD(boolean HD) {
+            Animation.this.HD = HD;
+        }
     }
 
     public class Get {
@@ -335,10 +358,6 @@ public class Animation {
 
         public double scale() {
             return scale;
-        }
-
-        public double branchOffsetAngle() {
-            return Math.toDegrees(branchOffsetAngle);
         }
 
         public double startLength() {
@@ -365,10 +384,6 @@ public class Animation {
             return endstroke;
         }
 
-        public int numStartBranches() {
-            return numStartBranches;
-        }
-
         public int numRepotitions() {
             return numRepotitions;
         }
@@ -383,6 +398,14 @@ public class Animation {
 
         public boolean hasRenderFrames() {
             return outputToList;
+        }
+
+        public boolean outputToList() {
+            return outputToList;
+        }
+
+        public int frameRate() {
+            return frameRate;
         }
     }
 }
